Single test module for random spanning tree
