+++ test/CMakeLists.txt
@@ -2,6 +2,7 @@ include_directories("${CMAKE_CURRENT_SOURCE_DIR}/../3rdparty")
 
 set(tests
   spanning_tree_test.cpp
+  random_spanning_tree_test.cpp
 )
 
 foreach(test ${tests})
